requirements.txt
# Core AI/ML Dependencies
openai>=1.12.0
pydantic>=2.6.0
agents>=0.1.0
mlflow>=2.10.0

# Database & Storage
chromadb>=0.4.22

# HTTP Clients
httpx>=0.26.0
requests>=2.31.0

# Search & Retrieval
tavily-python>=0.3.0

# Configuration & Environment
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2

# Data Processing Tools (optional, for data_processing/ directory)
# Uncomment if you need to run data processing scripts
# PyMuPDF>=1.23.0
# tiktoken>=0.5.2

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# mypy>=1.8.0
# black>=24.0.0
# ruff>=0.1.0
